initial_size -> min_size & small fixes

+++ src/cereggii/atomic_dict/node_ops.c
@@ -18,7 +18,7 @@ atomic_dict_compute_raw_node(atomic_dict_node *node, atomic_dict_meta *meta)
     assert(node->distance <= meta->max_distance);
     node->node =
         (node->index << (meta->node_size - meta->log_size))
-        | (meta->distance_mask - (node->distance << meta->tag_size))
+        | (meta->distance_mask - ((uint64_t) node->distance << (uint64_t) meta->tag_size))
         | (node->tag & meta->tag_mask);
 }
